backend/pyproject.toml
[project]
name = "spark"
version = "0.1.0"
description = "spark using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.130.0,<1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "aiohttp>=3.8.0",
    "moviepy>=1.0.3",
    "pillow>=10.0.0",
    "flask>=2.3.0",
    "gradio>=4.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
]

[project.scripts]
kickoff = "spark.main:kickoff"
run_crew = "spark.main:kickoff"
plot = "spark.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
